src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Maxim Tsaregradtsev <3m3rcy3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.11.3"
pytest = "^8.3.5"
pydantic-settings = "^2.8.1"
loguru = "^0.7.3"
pyyaml = "^6.0.2"
types-pyyaml = "^6.0.12.20250402"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
mypy = "^1.15.0"
mypy-extensions = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ['pydantic.mypy']
strict = true
ignore_missing_imports = true